Fixed SortingService.CurrentOption Bug
