--- GitTrends.Mobile.Shared/Constants/GitHubUserNotFoundConstants.cs
@@ -1,10 +1,0 @@
-﻿namespace GitTrends.Mobile.Shared
-{
-    public static class GitHubUserNotFoundConstants
-    {
-        public const string Title = "GitHub User Not Found";
-        public const string Description = "Sign in to GitHub on the Settings Page";
-        public const string Accept = "OK";
-        public const string Decline = "Cancel";
-    }
-}
