+++ GitTrends.Mobile.Shared/GitTrends.Mobile.Shared.projitems
@@ -19,7 +19,6 @@
     <Compile Include="$(MSBuildThisFileDirectory)Constants\AutomationIds\ReferringSitesPageAutomationIds.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Constants\AutomationIds\SplashScreenPageAutomationIds.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Constants\AutomationIds\TrendsPageAutomationIds.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Constants\GitHubUserNotFoundConstants.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Constants\GitHubLoginButtonConstants.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Constants\BackdoorMethodConstants.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Constants\DemoDataConstants.cs" />
